Refactored Helper Folder to Utility
